Fix git overwrite bug
Changes in git commits received by auto-pulling before an `add` would
be overwritten upon committing the add.  This occurred because `add`
did not reload the index after pulling and wrote the stale index as
the sum of its changes.

The same issue existed for the `rm` and `refresh` commands.

It has been fixed by always git-pulling *before* loading the index.